create outputs.py
