+++ corgisim/scene.py
@@ -6,7 +6,7 @@ from synphot import units, SourceSpectrum, SpectralElement, Observation
 from synphot.units import validate_wave_unit, convert_flux, VEGAMAG
 import cgisim
 import os
-from datetime import datetime, timezone, timedelta
+
 
 class Scene():
     ''' 
@@ -426,46 +426,6 @@ class SimulatedScene():
         '''
         pass
 
-    def save_hdu_to_fits(self, hdul, outdir=None, overwrite=True, write_as_L1=False, filename=None):
-        """
-        Save an Astropy HDUList to a FITS file.
-
-        Parameters:
-        - hdul (astropy.io.fits.HDUList): The HDUList object to be saved.
-        - outdir (str, optional): Output directory. Defaults to the current working directory.
-        - overwrite (bool): If True, overwrite the file if it already exists. Default is True.
-        - write_as_L1 (bool): If True, the file will be named according to the L1 naming convention.
-        - filename (str, optional): Name of the output FITS file (without ".fits" extension). 
-                                    Required if write_as_L1 is False.
-        """
-        if outdir is None:
-            outdir = os.getcwd()
-
-        os.makedirs(outdir, exist_ok=True)
-
-        # Handle naming logic
-        if write_as_L1:
-            #following the name convention for L1 product
-            if (hdul[1].data.shape[0] != 1200) or (hdul[1].data.shape[1] != 2200):
-                raise ValueError("Only full frame image can be save as L1 products")
-                
-            if filename is not None:
-                raise Warning("The provided filename is overridden for L1 products.")
-            prihdr = hdul[0].header
-            exthdr = hdul[1].header
-
-            time_in_name = isotime_to_yyyymmddThhmmsss(exthdr['FTIMEUTC'])
-            filename = f"CGI_{prihdr['VISITID']}_{time_in_name}_L1_"
-        else:
-            if filename is None:
-                raise ValueError("Filename must be provided when write_as_L1 is False.")
-
-        # Construct full file path
-        filepath = os.path.join(outdir, f"{filename}.fits")
-        
-        # Write the HDUList to file
-        hdul.writeto(filepath, overwrite=overwrite)
-        print(f"Saved FITS file to: {filepath}")
 
 
 
@@ -521,35 +481,4 @@ def is_valid_spectral_type(spectral_type):
     return bool(match)
 
 
-
-def isotime_to_yyyymmddThhmmsss(timestr):
-    """
-    Format an ISO time string into a custom format yyyymmddThhmmsss
-
-    Parameters
-    ----------
-    timestr : str
-        ISO time (e.g., '2025-04-25T23:18:04.786184+00:00').
-
-    Returns
-    -------
-    str
-        Time formatted as 'yyyymmddThhmmsss' (e.g., '20250425T2318048').
-    """
-    # Parse the input ISO format time
-    t = datetime.fromisoformat(timestr)
-
-    # Round to nearest 0.1 second
-    microsecond = t.microsecond
-    tenth_sec = round(microsecond / 1e5)  # how many tenths
-    t = t.replace(microsecond=0)  # reset microseconds
-
-    if tenth_sec == 10:
-        # carry over 1 second if rounding goes to 10
-        t += timedelta(seconds=1)
-        tenth_sec = 0
-
-    # Format as yyyymmddThhmmsss
-    out = t.strftime("%Y%m%dT%H%M%S") + str(tenth_sec)
-    return out
